Add files via upload
**Def:
The Naive Bayes classification algorithm is a probabilistic classifier. It is based on probability models that incorporate strong independence assumptions. The independence assumptions often do not have an impact on reality. Therefore they are considered naive.

I used this in one of my data mining projects and got accuracy of 87%.

**Source: https://www.ibm.com/support/knowledgecenter/SSEPGG_10.1.0/com.ibm.im.overview.doc/c_naive_bayes_classification.html